Improve reliability of specifying dataset
Constructing the URL manually was fragile. Instead, use the same
logic for generating the RDF IRI as is used by the RDF serializer
which DataOntoSearch uses when downloading the RDF. This way, they
should arrive at the same result (as long as the site URL stays the
same).

This also involved changed APIs at DataOntoSearch, so this brings
compatibility with tobinus/OTD-semantic-framework@ae3c706630ef74eccfa9c2915d96367ce52e0893